+++ BaggedTrees/gbt_train.cpp
@@ -153,10 +153,10 @@ int main(int argc, char* argv[])
 	int attrIds[attrN];       
 	fill_n(attrIds, attrN, 0); // initialize all attrIds 0:notused 1:used
 	int num = 0;
-	int num1 = 0;
+	double num1 = 0;
 
 	int* numUsed = &num; // number of used features
-	int* compN = &num1; // number of approximate operations for training the model
+	double* compN = &num1; // number of approximate operations for training the model
 
 	if(topAttrN == -1)
 		topAttrN = attrN;
