+++ shared/TreeNode.h
@@ -57,7 +57,7 @@ public:
 	void traverse(int itemNo, double coef, double& ltCoef, double& rtCoef, DATA_SET dset);
 
 	//splits the node; grows two offsprings 
-	bool split(double alpha, double rootVar, double* pEntropy = NULL, double mu = 0, int *attrIds = NULL, int s = 20, int* numUsed = NULL, int* compN = NULL);
+	bool split(double alpha, double rootVar, double* pEntropy = NULL, double mu = 0, int *attrIds = NULL, int s = 20, int* numUsed = NULL, double* compN = NULL);
 
 	//saves the node into a binary file
 	void save(fstream& fsave);
@@ -77,16 +77,16 @@ private:
 	void makeLeaf(double nodeMean); 
 
 	//finds and sets a splitting info with the best MSE
-	bool setSplit(double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, int *attrIds = NULL, int *numUsed = NULL, int* compN = NULL);
+	bool setSplit(double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, int *attrIds = NULL, int *numUsed = NULL, double* compN = NULL);
 
 	//finds and sets a splitting info with the best MSE when missing values present in the data
 	bool setSplitMV(double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, int *attrIds = NULL);
 
 	//finds and sets a splitting info with the approximate best MSE, using the idea of groupTest and binarySearch
-	bool setGroupSplit(double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, int *attrIds = NULL, int s = 20, int *numUsed = NULL, int* compN = NULL);
+	bool setGroupSplit(double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, int *attrIds = NULL, int s = 20, int *numUsed = NULL, double* compN = NULL);
 
 	// evaluate a single variable split or single group of variables split 
-	bool singleSplit(SplitInfov& bestSplits, double& bestEval, int attr, dipairv* pSortedVals, double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, int* compN = NULL);
+	bool singleSplit(SplitInfov& bestSplits, double& bestEval, int attr, dipairv* pSortedVals, double nodeV, double nodeSum, double squares, double rootVar, double mu = 0, double* compN = NULL);
 
 	//evaluates boolean split
 	double evalBool(SplitInfo& canSplit, double nodeV, double nodeSum, double squares, double rootVar);
